# hospital-server/docs/swagger/swagger.yaml
basePath: /
definitions:
...
  dto.CreateHospitalReq:
    properties:
      address:
        type: string
      bed_count:
        type: integer
      category_id:
        type: string
      city:
        type: string
      code:
        maxLength: 50
        type: string
      contact_email:
        type: string
      contact_name:
        type: string
      contact_phone:
        type: string
      department_count:
        type: integer
      fields:
        additionalProperties:
          type: string
        description: dynamic field key->value
        type: object
      is_specialized:
        type: boolean
      level:
        type: string
      name:
        maxLength: 200
        type: string
      owner_user_id:
        type: string
      province_id:
        type: string
      remark:
        type: string
      specialty_type:
        type: string
    required:
    - code
    - name
    type: object
...
  dto.LoginReq:
    properties:
      password:
        type: string
      username:
        type: string
    required:
    - password
    - username
    type: object
  dto.LoginResp:
    properties:
      expires_in:
        type: integer
      token:
        type: string
      user: {}
    type: object
  dto.PageResult:
    properties:
      list: {}
      page:
        type: integer
      page_size:
        type: integer
      total:
        type: integer
    type: object
  dto.Response:
    properties:
      code:
        type: integer
      data: {}
      message:
        type: string
    type: object
...
  dto.UpdateHospitalReq:
    properties:
      address:
        type: string
      bed_count:
        type: integer
      category_id:
        type: string
      city:
        type: string
      contact_email:
        type: string
      contact_name:
        type: string
      contact_phone:
        type: string
      department_count:
        type: integer
      fields:
        additionalProperties:
          type: string
        type: object
      is_specialized:
        type: boolean
      level:
        type: string
      name:
        type: string
      owner_user_id:
        type: string
      province_id:
        type: string
      remark:
        type: string
      specialty_type:
        type: string
      status:
        type: integer
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: 医院信息系统 API
  title: HIS API
  version: "1.0"
paths:
  /api/admin/v1/auth/refresh:
    post:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/dto.Response'
            - properties:
                data:
                  $ref: '#/definitions/dto.LoginResp'
              type: object
      security:
      - BearerAuth: []
      summary: 刷新 Token
      tags:
      - auth
  /api/admin/v1/hospitals:
    get:
      parameters:
      - description: 页码
        in: query
        name: page
        type: integer
      - description: 每页数量
        in: query
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/dto.Response'
            - properties:
                data:
                  $ref: '#/definitions/dto.PageResult'
              type: object
      security:
      - BearerAuth: []
      summary: 医院列表
      tags:
      - hospitals
    post:
      consumes:
      - application/json
      parameters:
      - description: 医院信息
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.CreateHospitalReq'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dto.Response'
      security:
      - BearerAuth: []
      summary: 新建医院
      tags:
      - hospitals
  /api/admin/v1/hospitals/{id}:
    delete:
      parameters:
      - description: 医院 ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.Response'
      security:
      - BearerAuth: []
      summary: 删除医院
      tags:
      - hospitals
    get:
      parameters:
      - description: 医院 ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.Response'
      security:
      - BearerAuth: []
      summary: 获取医院
      tags:
      - hospitals
    put:
      consumes:
      - application/json
      parameters:
      - description: 医院 ID
        in: path
        name: id
        required: true
        type: string
      - description: 医院信息
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateHospitalReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.Response'
      security:
      - BearerAuth: []
      summary: 更新医院
      tags:
      - hospitals
...
  /api/auth/login:
    post:
      consumes:
      - application/json
      parameters:
      - description: 登录信息
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.LoginReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/dto.Response'
            - properties:
                data:
                  $ref: '#/definitions/dto.LoginResp'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.Response'
      summary: 用户登录
      tags:
      - auth
...
securityDefinitions:
  BearerAuth:
    description: 'JWT Bearer token (format: Bearer {token})'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"